requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]==1.7.4
cryptography
pydantic-settings
python-dotenv
boto3
redis